Moved interfaces to separate files for cleaner code
