+++ client/repositoryInterfaces/ILocationRepository.cs
@@ -1,0 +1,10 @@
+﻿using client.models;
+
+namespace client.repositories
+{
+    internal interface ILocationRepository
+    {
+        Task<Location> GetLocationDetails(string locationId);
+        Task<List<Location>> SearchLocations(string query);
+    }
+}
