+++ client/repositoryInterfaces/IPostArchivedRepository.cs
@@ -1,0 +1,11 @@
+﻿using client.models;
+
+namespace client.repositories
+{
+    internal interface IPostArchivedRepository
+    {
+        bool addPostArchivedToDB(PostArchived postArchived);
+        List<PostArchived> getAll();
+        bool removePostArchivedFromDB(PostArchived postArchived);
+    }
+}
