+++ client/repositoryInterfaces/IPostReportedRepository.cs
@@ -1,0 +1,11 @@
+﻿using client.models;
+
+namespace client.repositories
+{
+    internal interface IPostReportedRepository
+    {
+        bool addReportedPostToDB(PostReported postReported);
+        List<PostReported> getAll();
+        bool removeReportedPostFromDB(PostReported postReported);
+    }
+}
