+++ client/repositoryInterfaces/IUserRepository.cs
@@ -1,0 +1,10 @@
+﻿using client.models;
+
+namespace client.repositories
+{
+    interface IUserRepository
+    {
+        List<User> getAllUsers();
+        User getUserById(Guid id);
+    }
+}
