+++ client/services/MainService.cs
@@ -1,5 +1,6 @@
 ﻿using client.models;
 using client.repositories;
+using client.servicesInterfaces;
 using System.Net.Http;
 using System.Windows.Forms;
