+++ client/servicesInterfaces/ILocationService.cs
@@ -1,0 +1,10 @@
+﻿using client.models;
+
+namespace client.servicesInterfaces
+{
+    internal interface ILocationService
+    {
+        Task<Location> GetLocationById(string locationId);
+        Task<List<Location>> SearchLocations(string querry);
+    }
+}
