+++ client/servicesInterfaces/IPostReportedService.cs
@@ -1,0 +1,11 @@
+﻿using client.models;
+
+namespace client.servicesInterfaces
+{
+    internal interface IPostReportedService
+    {
+        bool addPostReported(PostReported postReported);
+        List<PostReported> getPostReportedList();
+        bool removePostReported(PostReported postReported);
+    }
+}
