+++ client/servicesInterfaces/IPostSavedService.cs
@@ -1,0 +1,11 @@
+﻿using client.models;
+
+namespace client.servicesInterfaces
+{
+    internal interface IPostSavedService
+    {
+        bool addPostSaved(PostSaved postSaved);
+        List<PostSaved> getPostSavedList();
+        bool removePostSaved(PostSaved postSaved);
+    }
+}
