+++ client/servicesInterfaces/IUserService.cs
@@ -1,0 +1,10 @@
+﻿using client.models;
+
+namespace client.servicesInterfaces
+{
+    interface IUserService
+    {
+        List<User> getAllUsers();
+        User getUserById(Guid id);
+    }
+}
